refactor: feedback 반영
- AcceptanceTest 생성 및 공통 기능 분리 / 각 인수 테스트 상속
- 파라메터 변수명 변경 id -> lineId
- 섹션 추가 메소드 생성 (중복 제거)
