+++ src/main/java/nextstep/subway/line/application/SectionService.java
@@ -20,19 +20,19 @@ public class SectionService{
     }
 
     @Transactional
-    public LineResponse addSection(final Long id, final SectionAddRequest sectionAddRequest) {
+    public LineResponse addSection(final Long lineId, final SectionAddRequest sectionAddRequest) {
         final Station upStation = stationService.findById(sectionAddRequest.getUpStationId());
         final Station downStation = stationService.findById(sectionAddRequest.getDownStationId());
-        final Line line = lineService.findById(id);
+        final Line line = lineService.findById(lineId);
 
         line.addSection(Section.of(upStation, downStation, sectionAddRequest.getDistance()));
         return LineResponse.from(line);
     }
 
     @Transactional
-    public void deleteSection(final Long id, final Long stationId) {
+    public void deleteSection(final Long lineId, final Long stationId) {
         final Station station = stationService.findById(stationId);
-        final Line line = lineService.findById(id);
+        final Line line = lineService.findById(lineId);
 
         line.deleteSection(station);
     }
