+++ src/test/java/nextstep/subway/line/domain/SectionsTest.java
@@ -1,6 +1,7 @@
 package nextstep.subway.line.domain;
 
 import nextstep.subway.line.exception.LineException;
+import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -14,11 +15,8 @@ public class SectionsTest {
     @BeforeEach
     void setUp() {
         sections = new Sections();
-
-        Section section = Section.of(StationTest.강남역, StationTest.사당역, 10L);
-        sections.add(section);
+        섹션_추가(StationTest.강남역, StationTest.사당역, 10L);
     }
-
 
     @Test
     void 상하행_모두_존재하지_않을때_오류() {
@@ -52,9 +50,7 @@ public class SectionsTest {
     @Test
     void 하행_종착역_추가() {
         // given
-        Section newSection = Section.of(StationTest.사당역, StationTest.이수역, 10L);
-        // when
-        sections.add(newSection);
+        섹션_추가(StationTest.사당역, StationTest.이수역, 10L);
         // then
         assertThat(sections.contains(Section.of(StationTest.강남역, StationTest.사당역, 10L))).isTrue();
         assertThat(sections.contains(Section.of(StationTest.사당역, StationTest.이수역, 10L))).isTrue();
@@ -63,9 +59,7 @@ public class SectionsTest {
     @Test
     void 상행_종착역_추가() {
         // given
-        Section newSection = Section.of(StationTest.이수역, StationTest.강남역, 4L);
-        // when
-        sections.add(newSection);
+        섹션_추가(StationTest.이수역, StationTest.강남역, 4L);
         // then
         assertThat(sections.contains(Section.of(StationTest.강남역, StationTest.사당역, 10L))).isTrue();
         assertThat(sections.contains(Section.of(StationTest.이수역, StationTest.강남역, 4L))).isTrue();
@@ -74,9 +68,7 @@ public class SectionsTest {
     @Test
     void 중간역_추가() {
         // given
-        Section newSection = Section.of(StationTest.강남역, StationTest.이수역, 4L);
-        // when
-        sections.add(newSection);
+        섹션_추가(StationTest.강남역, StationTest.이수역, 4L);
         // then
         assertThat(sections.contains(Section.of(StationTest.강남역, StationTest.이수역, 4L))).isTrue();
         assertThat(sections.contains(Section.of(StationTest.이수역, StationTest.사당역, 6L))).isTrue();
@@ -92,8 +84,7 @@ public class SectionsTest {
     @Test
     void 섹션_종착역_삭제() {
         // given
-        Section newSection = Section.of(StationTest.강남역, StationTest.이수역, 4L);
-        sections.add(newSection);
+        섹션_추가(StationTest.강남역, StationTest.이수역, 4L);
         // when
         sections.delete(StationTest.강남역);
         // then
@@ -103,8 +94,7 @@ public class SectionsTest {
     @Test
     void 섹션_중간역_삭제() {
         // given
-        Section newSection = Section.of(StationTest.강남역, StationTest.이수역, 4L);
-        sections.add(newSection);
+        섹션_추가(StationTest.강남역, StationTest.이수역, 4L);
         // when
         sections.delete(StationTest.이수역);
         // then
@@ -114,8 +104,8 @@ public class SectionsTest {
     @Test
     void 상행선_종점_구하기() {
         // given
-        sections.add(new Section(StationTest.사당역, StationTest.이수역, 10L));
-        sections.add(new Section(StationTest.이수역, StationTest.서울역, 10L));
+        섹션_추가(StationTest.사당역, StationTest.이수역, 10L);
+        섹션_추가(StationTest.이수역, StationTest.서울역, 10L);
         // when
         // then
         assertThat(sections.finalUpStation()).isEqualTo(StationTest.강남역);
@@ -124,10 +114,14 @@ public class SectionsTest {
     @Test
     void 하행선_종점_구하기() {
         // given
-        sections.add(new Section(StationTest.사당역, StationTest.이수역, 10L));
-        sections.add(new Section(StationTest.이수역, StationTest.서울역, 10L));
+        섹션_추가(StationTest.사당역, StationTest.이수역, 10L);
+        섹션_추가(StationTest.이수역, StationTest.서울역, 10L);
         // when
         // then
         assertThat(sections.finalDownStation()).isEqualTo(StationTest.서울역);
+    }
+
+    private void 섹션_추가(final Station upStation, final Station downStation, final long distance) {
+        sections.add(Section.of(upStation, downStation, distance));
     }
 }
